+++ Classes/Tools/ScalerTool.cs
@@ -36,16 +36,17 @@ namespace ChineseCheckers.Classes
             image = new Transform2DEffect() { Source = source };
             image.TransformMatrix = Matrix3x2.CreateScale(scaleWidth, scaleHeight);
             return image;
-        }   
-        public static float Xpos(float x)
+        }
+      
+        public static double Xpos(double x)
+
         {
-            return (float)(x * scaleWidth);
+            return (double)(x * scaleWidth);
 
         }
-        public static float Ypos(float y)
+        public static double Ypos(double y)
         {
-            return (float)(y * scaleHeight);
-
+            return (double)(y * scaleHeight);
         }
     }
 }
